Rename directory.
