+++ debugging/coreclr/debugSessionManager.ts
