+++ debugging/coreclr/dockerDebugConfigurationProvider.ts
