+++ debugging/coreclr/dockerManager.ts
