+++ debugging/coreclr/netCoreProjectProvider.ts
