+++ debugging/coreclr/prereqManager.ts
