+++ debugging/coreclr/tempFileProvider.ts
